refactor : ReviewService.modifyReview() 일부 수정
- 매개변수로 ReviewDTO만 받도록 수정
- ReviewService.getReview()를 사용해서 저장되어있던 리뷰를 가져오도록 수정
- modifyReview()에서 ReviewDTO를 받기 때문에 기존 빌더패턴으로 주입하던 부분 삭제
- Review.initKeywords()이후에 addAllKeywords()를 사용해서 수정된 키워드 일괄적으로 적용될 수 있도록 수정

+++ src/main/java/com/example/demo/service/ReviewService.java
@@ -51,49 +51,23 @@ public class ReviewService {
         storeService.addReviewInSummary(review);
     }
 
-    public List<ReviewDTO> getReviews(String storeId, Pageable pageable) {
-        List<Review> reviews = reviewRepository.findByStoreId(pageable, storeId);
-        List<ReviewDTO> reviewDTOS = new ArrayList<>();
-        for (Review r : reviews)
-            reviewDTOS.add(new ReviewDTO(r));
-        return reviewDTOS;
-    }
     @Transactional
-    public void modifyReview(String storeId, Long reviewEntryNo, ReviewDTO reviewDTO) {
-        User user = userRepository.findById(reviewDTO.getReviewEntryNo())
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 유저입니다!"));
-
-        Review review = reviewRepository.findById(reviewEntryNo)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 리뷰입니다!"));
+    public void modifyReview(ReviewDTO reviewDTO) {
+        Review review = getReview(reviewDTO.getReviewEntryNo());
+        if (!review.getUser().getEmail().equals(reviewDTO.getUserEmail()))
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "올바른 유저가 아닙니다!");
 
         Review oldReview = new Review();
         oldReview.modifyReview(review);
 
-        keywordRepository.deleteByReview_ReviewEntryNo(review.getReviewEntryNo());
-        review.modifyReview(Review.builder()
-                .reviewEntryNo(reviewEntryNo)
-                .reviewContent(reviewDTO.getReviewContent())
-                .starCount(reviewDTO.getStarCount())
-                .storeId(storeId)
-                .user(user)
-                .keywords(new ArrayList<>())
-                .build());
-
         reviewDTO.removeSameKeyword();
-        if (!isUsableKeywordContents(reviewDTO.getKeywords()))
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "존재하지 않는 키워드입니다!");
+        review.modifyReview(reviewDTO);
 
-        for (String k : reviewDTO.getKeywords()) {
-            Keyword keyword = Keyword.builder()
-                    .review(review)
-                    .user(user)
-                    .storeId(storeId)
-                    .keywordContent(getKeywordContent(k))
-                    .build();
-            review.getKeywords().add(keyword);
-        }
+        keywordRepository.deleteByReview_ReviewEntryNo(review.getReviewEntryNo());
 
-        reviewRepository.save(review);
+        review.initKeywords();
+        review.addAllKeywords(keywordContentService.getAllKeywordContent(reviewDTO.getKeywords()));
+
         storeService.modifyReviewInSummary(review, oldReview);
     }
